requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
lxml==5.3.0
moviepy==2.0.0.dev2
natsort==8.4.0
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydub==0.25.1
python-docx==1.1.2
python-dotenv==1.0.1
rsa==4.9
soupsieve==2.6
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
Flask==2.3.3
openai==0.27.10
requests==2.31.0
gunicorn==21.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2